requirements.txt
streamlit
pandas
plotly
mysql-connector-python
requests
python-dotenv
openai